requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
fastapi==0.70.1
gcloud==0.17.0
googleapis-common-protos==1.54.0
gunicorn==20.1.0
h11==0.12.0
httplib2==0.20.2
httptools==0.3.0
idna==3.3
jws==0.1.3
oauth2client==3.0.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pydantic==1.8.2
pyparsing==3.0.6
Pyrebase==3.0.27
python-dotenv==0.19.2
python-firebase==1.2
python-jwt==2.0.1
python-multipart==0.0.5
PyYAML==6.0
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
